Enable pushing files for both android and iOS
