+++ src/Appium.Net/Appium/AppiumCommandExecutionHelper.cs
@@ -15,6 +15,7 @@
 using System;
 using OpenQA.Selenium.Appium.Interfaces;
 using System.Collections.Generic;
+using System.IO;
 using System.Text;
 using OpenQA.Selenium.Appium.Android;
 using OpenQA.Selenium.Appium.Enums;
@@ -120,6 +121,27 @@ namespace OpenQA.Selenium.Appium
         {
             var encodedContentBytes = Convert.FromBase64String(GetClipboard(executeMethod, ClipboardContentType.PlainText));
             return Encoding.UTF8.GetString(encodedContentBytes);
+        }
+
+        public static void PushFile(IExecuteMethod executeMethod, string pathOnDevice, byte[] base64Data) =>
+            executeMethod.Execute(AppiumDriverCommand.PushFile, new Dictionary<string, object>()
+                { ["path"] = pathOnDevice, ["data"] = base64Data });
+
+        public static void PushFile(IExecuteMethod executeMethod, string pathOnDevice, FileInfo file)
+        {
+            if (file == null)
+            {
+                throw new ArgumentException("The file argument should not be null");
+            }
+
+            if (!file.Exists)
+            {
+                throw new ArgumentException("The file " + file.FullName + " doesn't exist");
+            }
+
+            byte[] bytes = File.ReadAllBytes(file.FullName);
+            string fileBase64Data = Convert.ToBase64String(bytes);
+            PushFile(executeMethod, pathOnDevice, Convert.FromBase64String(fileBase64Data));
         }
 
         #endregion Device Commands
